requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.7
django-filter==2.1.0
django-import-export==1.2.0
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
jdcal==1.4
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
protobuf==3.7.1
PyMySQL==0.9.3
python-decouple==3.1
PyYAML==5.1
six==1.12.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
